Process info, files endpoint, started signatures
Process info: Now the api/v1/pids endpoint lists additional information about the running processes. Working as far as I can tell.
Files: Added an endpoint, api/v1/files, that exposes the root directory of the host and provides information on a file or directory based on the end of the URI. Working as far as I can tell.
Signatures: Determines whether or not a given binary has a Polyverse signature. Works correctly for a binary in a format recognized by objdump with a ".comment" section, that does not have a Polyverse signature. Have yet to test on a binary with a Polyverse signature. Doesn't yet work on binaries that objdump can't find a ".comment" section in.